Add dummy BCL reference assemblies for use in ApiDump
Only assemblies with types depended upon by the C# type system are
included, along with the compatibility shims for assemblies that include
these types across the various target frameworks. Assemblies are
generally kept as minimal as possible, but also include any attributes
we may want to display in ApiDump output in the future. The SNK files
used to sign the assemblies are taken from the dotnet/arcade repository.