Add a project to generate a test assembly
This will be used in future by a test suite. ApiDump will be run on it
and the output will be compared to expected output.

Some things are covered extensively but not quite exhaustively, such as
type parameter constraints. Other things are not yet covered at all,
such as operators, type conversions, and nested types. I suspect it
covers some things ApiDump may not even support properly yet, such as
default interface implementations in C# 8.