Don't enable ReadyToRun in the project file
The benefit to ApiDump is minimal. About a third of a second is shaved
off execution time which is not significant when assembly processing
time is usually at least a second (numbers based on testing done on the
Roslyn libraries). It does not seem worth the size increase when
publishing with --no-self-contained, which is slightly more than double.

For custom publishing scenarios, R2R can still be enabled via the
command line (-p:PublishReadyToRun=true).