Add a global.json file
This is being added to avoid automatically using prereleases SDKs.
No version is specified so we otherwise retain the default roll-forward
behaviour and permit any SDK version that supports the TFMs we use.

This should be modified to allow prerelease SDKs only if, when, and
while we target a prerelease framework version.