Use pattern-based null checking everywhere
For most types this is just a point of style and good practice, but for
types where there is a compatible == operator overload, using == and !=
for null checking will unnecessarily invoke the overload, whereas using
"is null" will always just emit the branch-if-zero instruction.

There is an outstanding TODO to adopt 'is not null' when C# 9 ships.