Move to .NET 5 and C# 9, address new warnings
This only updates the language version ApiDump itself uses, and doesn't
add any support for recognising C# 9 features in analysed assemblies or
using new syntax in the output. The one exception is the handling of
default constant values, which is updated to use "null" for function
pointers and "0" for native ints.